Get rid of CODEOWNERS file + consolidate tpu_commons/layers/common to tpu_commons/layers/jax
