--- .github/CODEOWNERS
@@ -1,2 +1,0 @@
-tpu_commons/layers/common/ @gpolovets1 @bzgoogle
-tpu_commons/layers/recipes/ @gpolovets1 @bzgoogle
