+++ tests/layers/jax/layers/attention/test_common_attention.py
@@ -8,9 +8,9 @@ from flax import nnx
 from jax.sharding import Mesh
 from parameterized import parameterized
 
-from tpu_commons.layers.common.attention.attention import Attention
 from tpu_commons.layers.common.attention_metadata import AttentionMetadata
-from tpu_commons.layers.jax.attention import get_kv_cache_shape
+from tpu_commons.layers.jax.attention.attention import Attention
+from tpu_commons.layers.jax.attention_interface import get_kv_cache_shape
 
 KVCache = Tuple[jax.Array, jax.Array]
