+++ tests/layers/jax/layers/attention/test_deepseek_v3_attention.py
@@ -7,9 +7,9 @@ from flax import nnx
 from jax.sharding import Mesh
 from parameterized import parameterized
 
-from tpu_commons.layers.common.attention.deepseek_v3_attention import MLA
 from tpu_commons.layers.common.attention_metadata import AttentionMetadata
-from tpu_commons.layers.jax.attention import get_kv_cache_shape
+from tpu_commons.layers.jax.attention.deepseek_v3_attention import MLA
+from tpu_commons.layers.jax.attention_interface import get_kv_cache_shape
 
 
 class TestMLA(unittest.TestCase):
