+++ tests/test_base.py
@@ -4,7 +4,7 @@ import warnings
 from dataclasses import dataclass, field, fields
 from typing import Any, List, Mapping
 
-from tpu_commons.layers.common.base import Config
+from tpu_commons.layers.jax.base import Config
 
 # Use the 'warnings' module to globally ignore warnings within this block
 vllm_logger = logging.getLogger("vllm")
