+++ tpu_commons/attention/backends/pallas_torchax.py
@@ -12,7 +12,7 @@ from vllm.attention.backends.abstract import (AttentionBackend, AttentionImpl,
 
 from tpu_commons import utils
 from tpu_commons.layers.common.attention_metadata import AttentionMetadata
-from tpu_commons.layers.jax.attention import attention
+from tpu_commons.layers.jax.attention_interface import attention
 from tpu_commons.logger import init_logger
 # Register custom op dispatcher.
 from tpu_commons.models.vllm.vllm_model_wrapper_context import \
