+++ tpu_commons/layers/jax/glossary.md
