+++ tpu_commons/models/jax/phi3.py
@@ -9,8 +9,8 @@ from vllm.config import VllmConfig
 
 from tpu_commons import utils
 from tpu_commons.layers.common.attention_metadata import AttentionMetadata
-from tpu_commons.layers.jax.attention import attention
-from tpu_commons.layers.jax.rope import apply_longrope, apply_rope
+from tpu_commons.layers.jax.attention_interface import attention
+from tpu_commons.layers.jax.rope_interface import apply_longrope, apply_rope
 from tpu_commons.logger import init_logger
 from tpu_commons.models.jax.utils.weight_utils import (MetadataMap,
                                                        load_hf_weights)
