+++ tpu_commons/models/jax/qwen2_5_vl.py
@@ -15,7 +15,7 @@ from vllm.config import VllmConfig
 
 from tpu_commons import utils as utils
 from tpu_commons.layers.common.attention_metadata import AttentionMetadata
-from tpu_commons.layers.jax.attention import sharded_flash_attention
+from tpu_commons.layers.jax.attention_interface import sharded_flash_attention
 from tpu_commons.logger import init_logger
 from tpu_commons.models.jax.qwen2 import Qwen2ForCausalLM
 # from vllm.model_executor.models.interfaces import MultiModalEmbeddings
